commit finalizacion README.md y presentación

+++ src/main/java/es/upm/miw/patronstrategy/Average.java
@@ -4,6 +4,6 @@ public class Average implements Strategy {
 
     @Override
     public int calculate(int firstNumber, int secondNumber) {
-        return (int) (firstNumber + secondNumber) / 2;
+        return (firstNumber + secondNumber) / 2;
     }
 }
